requirements.txt
requests
matplotlib
pandas
python-dotenv
psutil
coloredlogs
colorama
